% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTOSD_function.R
\name{TTOSD_function}
\alias{TTOSD_function}
\title{Run TTO Signal detection}
\usage{
TTOSD_function(
  dataset,
  Product_var,
  Product_val,
  Event_var,
  TTO_var,
  interval,
  min
)
}
\arguments{
\item{dataset}{Input dataset containing at least a column for the product, a column for the event and a column for the TTO (in days)}

\item{Product_var}{Name of the field containing the Product identifier}

\item{Product_val}{List of products for which the TTO Signal detection will run (e.g. c("ZOSTER (SHINGRIX)","ZOSTER (ZOSTAVAX)"))}

\item{Event_var}{Name of the field containing the Event identifier}

\item{TTO_var}{Name of the field containing the TTO information (in days). Needs to be numeric}

\item{interval}{Length (in days) of the interval for the TTO Signal detection}

\item{min}{Minimal number of rows with non-missing TTO information in the user-defined interval for performing TTO Signal Detection}
}
\value{

}
\description{
Run TTO Signal detection
}
