% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTOSD_function.R
\name{TTOSD_function}
\alias{TTOSD_function}
\title{Title}
\usage{
TTOSD_function(
  dataset,
  Product_var,
  Product_val,
  Event_var,
  TTO_var,
  interval,
  min
)
}
\arguments{
\item{min}{}
}
\value{

}
\description{
Title
}
